// needrelay-frontend/src/api/generated/models/needPriority.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NeedRelay API
 * Crisis and emergency resource coordination API
 * OpenAPI spec version: 0.1.0
 */

export type NeedPriority = typeof NeedPriority[keyof typeof NeedPriority];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NeedPriority = {
  CRITICAL: 'CRITICAL',
  HIGH: 'HIGH',
  NORMAL: 'NORMAL',
  LOW: 'LOW',
} as const;